stDIKO is now a representation of your stake in the pool and auto-compounds rewards
