+++ clarity/contracts/arkadiko-stake-registry-v1-1.clar
@@ -70,8 +70,7 @@
     )
       (asserts! (is-eq (as-contract tx-sender) (contract-of registry-trait)) ERR-WRONG-REGISTRY)
       (asserts! (is-eq (get deactivated-block pool-info) u0) ERR-POOL-INACTIVE)
-      (try! (contract-call? pool-trait stake registry-trait token-trait tx-sender amount))
-      (ok amount)
+      (contract-call? pool-trait stake registry-trait token-trait tx-sender amount)
     )
   )
 )
@@ -84,8 +83,7 @@
       (pool-info (unwrap! (map-get? pools-data-map { pool: pool }) ERR-POOL-EXIST))
     )
       (asserts! (is-eq (as-contract tx-sender) (contract-of registry-trait)) ERR-WRONG-REGISTRY)
-      (try! (contract-call? pool-trait unstake registry-trait token-trait tx-sender amount))
-      (ok amount)
+      (contract-call? pool-trait unstake registry-trait token-trait tx-sender amount)
     )
   )
 )
