+++ clarity/contracts/tests/arkadiko-stake-pool-diko-xusd-tv1-1.clar
@@ -1,4 +1,4 @@
-;; Stake Pool - Stake DIKO to get sDIKO
+;; Stake Pool - Stake DIKO-xUSD LP tokens
 ;; 
 ;; A fixed amount of rewards per block will be distributed across all stakers, according to their size in the pool
 ;; Rewards will be automatically staked before staking or unstaking. 
@@ -18,8 +18,7 @@
 (define-constant ERR-WRONG-REGISTRY (err u18004))
 
 ;; Constants
-(define-constant POOL-TOKEN .arkadiko-token)
-(define-constant BLOCKS-PER-YEAR u52560)
+(define-constant POOL-TOKEN .arkadiko-swap-token-diko-xusd)
 
 ;; Variables
 (define-data-var total-staked uint u0)
@@ -94,16 +93,13 @@
       ;; Update cumm reward per stake now that total is updated
       (unwrap-panic (increase-cumm-reward-per-stake registry-trait))
 
-      ;; Mint stDIKO
-      (try! (contract-call? .arkadiko-dao mint-token .stdiko-token amount staker))
-
-      ;; Transfer DIKO to this contract
-      (try! (contract-call? .arkadiko-token transfer amount staker (as-contract tx-sender) none))
+      ;; Transfer LP token to this contract
+      (try! (contract-call? .arkadiko-swap-token-diko-xusd transfer amount staker (as-contract tx-sender) none))
 
       ;; Update sender stake info
       (map-set stakes { staker: staker } { uamount: new-stake-amount, cumm-reward-per-stake: (var-get cumm-reward-per-stake) })
 
-      (ok new-stake-amount)
+      (ok amount)
     )
   )
 )
@@ -134,16 +130,13 @@
       ;; Update cumm reward per stake now that total is updated
       (unwrap-panic (increase-cumm-reward-per-stake registry-trait))
 
-      ;; Burn stDIKO 
-      (try! (contract-call? .arkadiko-dao burn-token .stdiko-token amount staker))
-
-      ;; Transfer DIKO back from this contract to the user
-      (try! (contract-call? .arkadiko-token transfer amount (as-contract tx-sender) staker none))
+      ;; Transfer LP token back from this contract to the user
+      (try! (contract-call? .arkadiko-swap-token-diko-xusd transfer amount (as-contract tx-sender) staker none))
 
       ;; Update sender stake info
       (map-set stakes { staker: staker } { uamount: new-stake-amount, cumm-reward-per-stake: (var-get cumm-reward-per-stake) })
 
-      (ok new-stake-amount)
+      (ok amount)
     )
   )
 )
@@ -151,30 +144,24 @@
 ;; Sender can unstake all tokens without claiming rewards
 (define-public (emergency-withdraw (registry-trait <stake-registry-trait>))
   (begin
-    ;; Save currrent cumm reward per stake
-    (unwrap-panic (increase-cumm-reward-per-stake registry-trait))
-
     ;; Check given registry
     (asserts! (is-eq (contract-of registry-trait) (unwrap-panic (contract-call? .arkadiko-dao get-qualified-name-by-name "stake-registry"))) ERR-WRONG-REGISTRY)
 
-    (let (
-      ;; Calculate new stake amount
-      (amount (unwrap-panic (contract-call? .stdiko-token get-balance tx-sender)))
+    ;; Save currrent cumm reward per stake
+    (unwrap-panic (increase-cumm-reward-per-stake registry-trait))
 
+    (let (
       (stake-amount (get-stake-amount-of tx-sender))
-      (new-stake-amount (- stake-amount amount))
+      (new-stake-amount u0)
     )
       ;; Update total stake
-      (var-set total-staked (- (var-get total-staked) amount))
+      (var-set total-staked (- (var-get total-staked) stake-amount))
 
       ;; Update cumm reward per stake now that total is updated
       (unwrap-panic (increase-cumm-reward-per-stake registry-trait))
 
-      ;; Burn stDIKO 
-      (try! (contract-call? .arkadiko-dao burn-token .stdiko-token amount tx-sender))
-
-      ;; Transfer DIKO back from this contract to the user
-      (try! (contract-call? .arkadiko-token transfer amount (as-contract tx-sender) tx-sender none))
+      ;; Transfer LP token back from this contract to the user
+      (try! (contract-call? .arkadiko-swap-token-diko-xusd transfer stake-amount (as-contract tx-sender) tx-sender none))
 
       ;; Update sender stake info
       (map-set stakes { staker: tx-sender } { uamount: new-stake-amount, cumm-reward-per-stake: (var-get cumm-reward-per-stake) })
@@ -186,7 +173,7 @@
 
 (define-public (get-apy-for (registry-trait <stake-registry-trait>) (staker principal))
   (let (
-    (rewards-per-block (unwrap-panic (contract-call? registry-trait get-rewards-per-block-for-pool .arkadiko-stake-pool-diko-v1-1)))
+    (rewards-per-block (unwrap-panic (contract-call? registry-trait get-rewards-per-block-for-pool .arkadiko-stake-pool-diko-xusd-v1-1)))
     (diko-staked (get-stake-amount-of staker))
     (reward-percentage (/ u100 (/ (var-get total-staked) diko-staked)))
     (diko-per-year (* rewards-per-block reward-percentage))
@@ -251,7 +238,7 @@
 ;; Calculate current cumm reward per stake
 (define-public (calculate-cumm-reward-per-stake (registry-trait <stake-registry-trait>))
   (let (
-    (rewards-per-block (unwrap-panic (contract-call? registry-trait get-rewards-per-block-for-pool .arkadiko-stake-pool-diko-v1-1)))
+    (rewards-per-block (unwrap-panic (contract-call? registry-trait get-rewards-per-block-for-pool .arkadiko-stake-pool-diko-xusd-v1-1)))
     (current-total-staked (var-get total-staked))
     (last-block-height (get-last-block-height registry-trait))
     (block-diff (- last-block-height (var-get last-reward-increase-block)))
@@ -274,7 +261,7 @@
 ;; Return current block height, or block height when pool was deactivated
 (define-private (get-last-block-height (registry-trait <stake-registry-trait>))
   (let (
-    (deactivated-block (unwrap-panic (contract-call? registry-trait get-pool-deactivated-block .arkadiko-stake-pool-diko-v1-1)))
+    (deactivated-block (unwrap-panic (contract-call? registry-trait get-pool-deactivated-block .arkadiko-stake-pool-diko-xusd-v1-1)))
     (pool-active (is-eq deactivated-block u0))
   )
     (if (is-eq pool-active true)
@@ -286,5 +273,5 @@
 
 ;; Initialize the contract
 (begin
-  (var-set last-reward-increase-block u2000)
+  (var-set last-reward-increase-block block-height)
 )
